% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_relation.R
\name{causal_relation}
\alias{causal_relation}
\title{Identify the Presence of Causal Relationships}
\usage{
causal_relation(context,
                variable_list,
                LLM_model = "gpt-4o",
                max_tokens = 2000,
                update_key = FALSE)
}
\arguments{
\item{context}{A character vector specifying the context for which a theory should be developed. If it is not feasible to identify a particular context, the argument can be set to NULL.}

\item{variable_list}{A vector containing all variables that need to be included in the theory.}

\item{LLM_model}{The LLM model that should be used to generate output: \code{"gpt-4o"} (default), \code{"gpt-4"}, \code{"gpt-4-turbo"} \code{"gpt-3.5-turbo"}, \code{"llama-3"} (specifically points to Llama-3-70B-Chat-hf, accessed via Hugging Face), or \code{"mixtral"} (specifically points to Mixtral-8x7B-Instruct-v0.1, accessed via Together.ai).}

\item{max_tokens}{The maximum number of tokens the LLM should generate. Be careful when adjusting this argument. Reducing the maximum token limit will reduce the cost but may result in incomplete answers. Conversely, increasing the token limit can be advantageous for obtaining more detailed responses. The maximum number of tokens depends on the model (\code{6000} for \code{"gpt-4o"}, \code{"gpt-4"}, and \code{"gpt-4-turbo"}, \code{3000} for \code{"gpt-3.5-turbo"}, and \code{2000} for \code{"mixtral"}).}

\item{update_key}{If \code{update_key = TRUE}, the function will prompt the user for a new API key and update the saved key. If \code{update_key = FALSE} (default), the function will use the existing API key if available.}
}
\value{
\itemize{
  \item \code{raw_LLM}: A dataframe containing the unprocessed LLM output along with some other LLM information, including:
    \itemize{
      \item \code{relationship}: Which variable pair.
      \item \code{iteration}: Iteration number.
      \item \code{LLM_model}: LLM model used.
      \item \code{prompt}: Prompt used.
      \item \code{system_prompt}: System prompt used.
      \item \code{content}: Unprocessed LLM output.
      \item \code{finish_reason}: Reason the LLM stopped generating output.
      \item \code{prompt_tokens}: Number of tokens used for the LLM prompt.
      \item \code{answer_tokens}: Number of tokens used for the LLM answer.
      \item \code{total_tokens}: Total number of tokens used.
      \item \code{error}: Error message, if any occurred.
    }
} \cr
\itemize{
  \item \code{relation_df}: A dataframe with three columns:
    \itemize{
      \item \code{var1}: Variable 1 of a unique variable pair.
      \item \code{var2}: Variable 2 of a unique variable pair.
      \item \code{prob_causal}: Probability of the presence of a causal relationship between var1 and var2.
    }
}
}
\description{
\code{causal_relation()} will generate all possible unique pairs of variables (e.g. 4 variables will give 5 * (5 - 1) / 2 = 10 unique pairs). It will then iterate through all these pairs and ask a Large Language Model (LLM) to indicate whether there is a causal or non-causal relationship between the variables.
}
\details{
To create a fully fledged theory from scratch, the functions in this R-packaged should be used in the following order:

\code{\link{var_list}} --> \code{\link{causal_relation}} --> \code{\link{causal_direction}} --> \code{\link{causal_sign}} --> \code{\link{cld_plot}}
}
\note{
The function and its output should be approached with caution. Depending on the specific LLM used, there may be financial implications. Furthermore, we wish to emphasise that the answers generated by an LLM should not be taken as absolute truth.
}
\examples{
\dontrun{
## Example input (context = "addiction")
data("vars")
vars$final_list

#---------------------------------------------------------------------------
## Default
# For a readily available, pre-made output example see: data("rel")
rel <- causal_relation(context = "addiction",
                       variable_list = vars$final_list)

# Check output
rel$relation_df
}
}
\references{
\url{https://platform.openai.com}
}
\seealso{
\code{\link{cld}},
\code{\link{var_list}},
\code{\link{causal_direction}},
\code{\link{causal_sign}},
\code{\link{cld_plot}},
}
\author{
Meike Waaijers
}
