% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cld.R
\name{cld}
\alias{cld}
\title{Quickly create a CLD using a LLM}
\usage{
cld(topic,
    variable_list,
    plot = TRUE,
    LLM_model = "gpt-4o",
    max_tokens = 2000,
    update_key = FALSE)
}
\arguments{
\item{topic}{A character vector specifying the topic for which a theory should be developed. If it is not feasible to identify a particular topic, the parameter can be set to NULL.}

\item{variable_list}{A vector containing all variables that need to be included in the theory.}

\item{plot}{If \code{plot = TRUE} (default), the function will generate network plot(s) visualizing the edge list(s).}

\item{LLM_model}{The LLM model that should be used to generate output: \code{"gpt-4o"} (default), \code{"gpt-4"}, \code{"gpt-4-turbo"} \code{"gpt-3.5-turbo"}, \code{"llama-3"} (specifically points to Llama-3-70B-Chat-hf, accessed via Hugging Face), or \code{"mixtral"} (specifically points to Mixtral-8x7B-Instruct-v0.1, accessed via Together.ai).}

\item{max_tokens}{The maximum number of tokens the LLM should generate. Be careful when adjusting this argument. Reducing the maximum token limit will reduce the cost but may result in incomplete answers. Conversely, increasing the token limit can be advantageous for obtaining more detailed responses. The maximum number of tokens depends on the model (\code{6000} for \code{"gpt-4o"}, \code{"gpt-4"}, and \code{"gpt-4-turbo"}, \code{3000} for \code{"gpt-3.5-turbo"}, and \code{2000} for \code{"mixtral"}).}

\item{update_key}{If \code{update_key = TRUE}, the function will prompt the user for a new API key and update the saved key. If \code{update_key = FALSE} (default), the function will use the existing API key if available.}
}
\value{
\itemize{
  \item \code{raw_LLM}: A dataframe containing the unprocessed LLM output along with some other LLM information, including:
    \itemize{
      \item \code{which_fun}: Which function the raw output comes from.
      \item \code{var}: Which variable is put in the prompt as cause variable. (NA for causal_relation function)
      \item \code{relationship}: Which variable pair.
      \item \code{iteration}: Iteration number.
      \item \code{LLM_model}: LLM model used.
      \item \code{prompt}: Prompt used.
      \item \code{system_prompt}: System prompt used.
      \item \code{content}: Unprocessed LLM output.
      \item \code{finish_reason}: Reason the LLM stopped generating output.
      \item \code{prompt_tokens}: Number of tokens used for the LLM prompt.
      \item \code{answer_tokens}: Number of tokens used for the LLM answer.
      \item \code{total_tokens}: Total number of tokens used.
      \item \code{error}: Error message, if any occurred.
    }
} \cr
\itemize{
  \item \code{sign_df}: A dataframe with nine columns:
    \itemize{
      \item \code{var1}: Variable 1 of a unique variable pair.
      \item \code{var2}: Variable 2 of a unique variable pair.
      \item \code{prob_causal}: Probability of the presence of a causal relationship between var1 and var2.
      \item \code{prob_var1_cause}: Probability of var1 being a cause variable.
      \item \code{prob_var1_pos}: Probability of a positive relationship where var1 is the cause.
      \item \code{prob_var1_neg}: Probability of a negative relationship where var1 is the cause.
      \item \code{prob_var2_cause}: Probability of var2 being a cause variable.
      \item \code{prob_var2_pos}: Probability of a positive relationship where var1 is the cause.
      \item \code{prob_var2_neg}: Probability of a negative relationship where var1 is the cause.
    }
} \cr
\itemize{
  \item
A plot and an edge list.
An edge list is a list containing all the edges between nodes. These nodes symbolise variables, while the edges denote the causal relationships that link these variables.

The specific outputted edge list is \code{dir_signs_edge_list}.
This is an edge list for a direction probability dataframe with sign indication (for more information about different edge lists see \code{\link{cld_plot}}).
\cr
\cr
\code{edge_list}: An edge list with 4 columns:
\itemize{
      \item \code{from}: Cause variable.
      \item \code{to}: Dependent variable.
      \item \code{weight}: Probability of a causal relationship between the Cause and Dependent variable.
      \item \code{sign}: Sign of the causal relationship (can be either "Positive", "Negative", or "Uncertain").
    }
}
}
\description{
The \code{cld()} function can be used to quickly generate a Causal Loop Diagram (CLD). This function uses the \code{\link{causal_relation}}, \code{\link{causal_direction}}, \code{\link{causal_sign}}, and \code{\link{cld_plot}} functions included in this R package. For greater control and access to more outputs, we recommend using these functions individually.
}
\details{
To create a theory from scratch, the functions in this R-package should be used in the following order:

\code{\link{var_list}} --> \code{\link{causal_relation}} --> \code{\link{causal_direction}} --> \code{\link{causal_sign}} --> \code{\link{cld_plot}} --> \code{\link{find_source}}
}
\note{
The function and its output should be approached with caution. Depending on the specific LLM used, there may be financial implications. Furthermore, we wish to emphasise that the answers generated by an LLM should not be taken as absolute truth.
}
\examples{
\dontrun{
## Example input (topic = "addiction")
data("vars")
vars$final_list

#---------------------------------------------------------------------------
## Default
# For a readily available, pre-made output example see: data("cld_example")

cld_example <- cld(topic = "addiction",
                   variable_list = vars$final_list)

# Check output
cld_example$sign_df
cld_example$edge_list
}

}
\references{
\url{https://platform.openai.com}
}
\seealso{
\code{\link{var_list}},
\code{\link{causal_relation}},,
\code{\link{causal_direction}},
\code{\link{causal_sign}},
\code{\link{cld_plot}},
\code{\link{find_source}}
}
\author{
Meike Waaijers
}
