% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cld_example.R
\docType{data}
\name{cld_example}
\alias{cld_example}
\title{cld_example data}
\format{
A list containing the following components:
\describe{
\itemize{
    \item \code{raw_LLM}: A dataframe containing the unprocessed LLM output along with some other LLM information, including:
    \describe{
    \itemize{
      \item \code{which_fun}: Which function the raw output comes from.
      \item \code{var}: Which variable is put in the prompt as cause variable. (NA for causal_relation function)
      \item \code{relationship}: Which variable pair.
      \item \code{iteration}: Iteration number.
      \item \code{LLM_model}: LLM model used.
      \item \code{prompt}: Prompt used.
      \item \code{system_prompt}: System prompt used.
      \item \code{content}: Unprocessed LLM output.
      \item \code{finish_reason}: Reason the LLM stopped generating output.
      \item \code{prompt_tokens}: Number of tokens used for the LLM prompt.
      \item \code{answer_tokens}: Number of tokens used for the LLM answer.
      \item \code{total_tokens}: Total number of tokens used.
      \item \code{error}: Error message, if any occurred.
    }} \cr
    \item \code{sign_df}: A dataframe with nine columns representing the probability of the presence, the direction, and the sign of causal relationships between various variable.
    \describe{
    \itemize{
      \item \code{var1}: Variable 1 of a unique variable pair.
      \item \code{var2}: Variable 2 of a unique variable pair.
      \item \code{prob_causal}: Probability of the presence of a causal relationship between var1 and var2.
      \item \code{prob_var1_cause}: Probability of var1 being a cause variable.
      \item \code{prob_var1_pos}: Probability of a positive relationship where variable 1 is the cause.
      \item \code{prob_var1_neg}: Probability of a negative relationship where variable 1 is the cause.
      \item \code{prob_var2_cause}: Probability of var2 being a cause variable.
      \item \code{prob_var2_pos}: Probability of a positive relationship where variable 2 is the cause.
      \item \code{prob_var2_neg}: Probability of a negative relationship where variable 2 is the cause.
    }} \cr
  \item \code{edge_list}: An edge list for a \code{dir_sign_df} input.
    \describe{
    \itemize{
      \item \code{from}: Cause variable.
      \item \code{to}: Dependent variable.
      \item \code{weight}: Probability of a causal relationship between the Cause and Dependent variable.
      \item \code{sign}: Sign of the causal relationship (either Positive, Negative or Uncertain).
    }}
}
}
}
\source{
Generated using the OpenAI platform: \url{https://platform.openai.com}
}
\usage{
cld_example
}
\description{
This data set contains an edge lists related to the context of addiction.
The data was generated using the \code{\link{cld}} function and the "gpt-4o" Large Language Model (LLM).
}
\examples{
data("cld_example")
}
\keyword{datasets}
