% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge_lists.R
\docType{data}
\name{edge_lists}
\alias{edge_lists}
\title{edge_lists data}
\format{
A list containing the following components:
\describe{
\itemize{
  \item \code{rel_edge_list}: Edge list for a relation probability dataframe (\code{relation_df}) input.
    \describe{
    \itemize{
      \item \code{from}: The "from" node, otherwise known as the "cause" variable.
      \item \code{to}: The "to" node, otherwise known as the "dependent" variable.
      \item \code{weight}: The weights associated with the causal relationship. In this case, the weights refer to the probability of there being a causal relationship.
    }} \cr
  \item \code{dir_edge_list}: Edge list for a direction probability dataframe (\code{direction_df}) input.
    \describe{
    \itemize{
      \item \code{from}: The "from" node, otherwise known as the "cause" variable.
      \item \code{to}: The "to" node, otherwise known as the "dependent" variable.
      \item \code{weight}: The weights associated with the causal relationship. In this case, the weights refer to the probability of there being a causal relationship.
    }} \cr
  \item \code{rel_sign_edge_list}: Edge list for a relation probability dataframe with sign indication (relation \code{sign_df}) input.
    \describe{
    \itemize{
      \item \code{from}: The "from" node, otherwise known as the "cause" variable.
      \item \code{to}: The "to" node, otherwise known as the "dependent" variable.
      \item \code{weight}: The weights associated with the causal relationship. In this case, the weights refer to the probability of there being a causal relationship.
      \item \code{sign}: The sign of the causal relationship (can be either "positive", "negative", or "uncertain").
    }} \cr
  \item \code{dir_sign_edge_list}: Edge list for a direction probability dataframe with sign indication (direction \code{sign_df}) input.
    \describe{
    \itemize{
      \item \code{from}: The "from" node, otherwise known as the "cause" variable.
      \item \code{to}: The "to" node, otherwise known as the "dependent" variable.
      \item \code{weight}: The weights associated with the causal relationship. In this case, the weights refer to the probability of there being a causal relationship.
      \item \code{sign}: The sign of the causal relationship (can be either "positive", "negative", or "uncertain").
    }}
}
}
}
\source{
Generated using the OpenAI platform: \url{https://platform.openai.com}
}
\usage{
edge_lists
}
\description{
This data set contains a list of 4 edge lists related to the topic of addiction.
The data was generated using the \code{\link{cld_plot}} function and the "gpt-4o" Large Language Model (LLM).
}
\examples{
data("edge_lists")
}
\keyword{datasets}
