% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rel.R
\docType{data}
\name{rel}
\alias{rel}
\title{rel data}
\format{
A list containing the following components:
\describe{
\itemize{
  \item \code{raw_LLM}: A dataframe containing the unprocessed LLM output along with some other LLM information, including:
    \describe{
    \itemize{
      \item \code{relationship}: Which variable pair.
      \item \code{iteration}: Iteration number.
      \item \code{LLM_model}: LLM model used.
      \item \code{prompt}: Prompt used.
      \item \code{system_prompt}: System prompt used.
      \item \code{content}: Unprocessed LLM output.
      \item \code{finish_reason}: Reason the LLM stopped generating output.
      \item \code{prompt_tokens}: Number of tokens used for the LLM prompt.
      \item \code{answer_tokens}: Number of tokens used for the LLM answer.
      \item \code{total_tokens}: Total number of tokens used.
      \item \code{error}: Error message, if any occurred.
    }} \cr
  \item \code{relation_df}: A dataframe with three columns representing the probability of the presence of causal relationships between various variable.
    \describe{
    \itemize{
      \item \code{var1}: Variable 1 of a unique variable pair.
      \item \code{var2}: Variable 2 of a unique variable pair.
      \item \code{prob_causal}: Probability of the presence of a causal relationship between var1 and var2.
    }}
}
}
}
\source{
Generated using the OpenAI platform: \url{https://platform.openai.com}
}
\usage{
rel
}
\description{
This data set contains the probabilities of the presence of
causal relationships between various variables related to the topic of addiction.
The data was generated using the \code{\link{causal_relation}} function and the "gpt-4o" Large Language Model (LLM).
}
\examples{
data("rel")
}
\keyword{datasets}
